requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.0.0=pypi_0
anyio=3.5.0=pypi_0
appnope=0.1.2=py38h10201cd_2
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38hea4295b_1
asttokens=2.0.5=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
augmentor=0.2.9=pypi_0
babel=2.9.1=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
bandit=1.6.0=py_1
beautifulsoup4=4.10.0=pyha770c72_0
beeprint=2.4.10=pypi_0
bleach=4.1.0=pyhd8ed1ab_0
c-ares=1.18.1=h3422bc3_0
ca-certificates=2021.10.8=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.0.0=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=py38hc67bbb8_0
charset-normalizer=2.0.12=pypi_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py38h6f2b01f_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.4=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd8ed1ab_0
flake8=4.0.1=pyhd8ed1ab_2
flatbuffers=2.0=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
focal-loss=0.0.2=pypi_0
fonttools=4.30.0=pypi_0
future=0.18.2=pypi_0
gast=0.5.3=pypi_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
google-auth=2.6.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
googleapis-common-protos=1.55.0=pypi_0
grpcio=1.44.0=py38h740c240_0
h5py=3.6.0=nompi_py38hacf61ce_100
hdf5=1.12.1=nompi_hd9dbc9e_104
idna=3.3=pypi_0
importlib-metadata=4.11.3=py38h10201cd_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.9.2=py38h2cb4d76_0
ipython=8.1.1=py38h10201cd_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.5=pyhd8ed1ab_0
jedi=0.18.1=py38h10201cd_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
json5=0.9.6=pypi_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter=1.0.0=py38h10201cd_7
jupyter-server=1.15.5=pypi_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_core=4.9.2=py38h10201cd_0
jupyterlab=3.3.2=pypi_0
jupyterlab-server=2.11.0=pypi_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.0=pypi_0
krb5=1.19.3=he492e65_0
libblas=3.9.0=13_osxarm64_openblas
libcblas=3.9.0=13_osxarm64_openblas
libclang=13.0.0=pypi_0
libcurl=7.82.0=h9476f53_0
libcxx=13.0.1=h6a5c8ee_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
liblapack=3.9.0=13_osxarm64_openblas
libnghttp2=1.47.0=hf30690b_0
libopenblas=0.3.18=openmp_h5dd58f0_0
libsodium=1.0.18=h27ca646_1
libssh2=1.10.0=h7a5bd25_2
libzlib=1.2.11=hee7b306_1013
llvm-openmp=13.0.1=h455960f_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=py38h33210d7_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mccabe=0.6.1=py_1
mistune=0.8.4=py38hea4295b_1005
nbclassic=0.3.7=pypi_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.4=py38h10201cd_0
nbformat=5.2.0=pyhd8ed1ab_0
ncurses=6.3=hc470f4d_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
notebook=6.4.10=pyha770c72_0
notebook-shim=0.1.0=pypi_0
numpy=1.21.5=py38hb29071a_0
oauthlib=3.2.0=pypi_0
openssl=3.0.0=h3422bc3_2
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=py38h3777fb4_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pbr=5.8.1=pyhd8ed1ab_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=pypi_0
pip=22.0.4=pyhd8ed1ab_0
prometheus_client=0.13.1=pyhd8ed1ab_0
promise=2.3=pypi_0
prompt-toolkit=3.0.27=pyha770c72_0
prompt_toolkit=3.0.27=hd8ed1ab_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=py38hea4295b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.8.0=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pydocstyle=6.1.1=pyhd8ed1ab_0
pyflakes=2.4.0=pyhd8ed1ab_0
pygments=2.11.2=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyrsistent=0.18.1=py38hea4295b_0
python=3.8.12=hd949e87_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd8ed1ab_0
pyyaml=6.0=py38hea4295b_3
pyzmq=22.3.0=py38h51b17a6_1
readline=8.1=hedafd6a_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=60.9.3=py38h10201cd_0
six=1.15.0=pypi_0
sklearn=0.0=pypi_0
smmap=3.0.5=pyh44b312d_0
sniffio=1.2.0=pypi_0
snowballstemmer=2.2.0=pyhd8ed1ab_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.37.1=h7e3ccbd_0
stack_data=0.2.0=pyhd8ed1ab_0
stevedore=3.5.0=py38h10201cd_2
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-datasets=4.5.2=pypi_0
tensorflow-deps=2.8.0=0
tensorflow-macos=2.8.0=pypi_0
tensorflow-metadata=1.7.0=pypi_0
tensorflow-metal=0.4.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=py38h10201cd_0
testpath=0.6.0=pyhd8ed1ab_0
tf-estimator-nightly=2.8.0.dev2021122109=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=he1e0b03_0
tornado=6.1=py38hea4295b_2
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.9=pypi_0
urwid=2.1.2=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.3.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.5.2=py38h10201cd_1
wrapt=1.14.0=pypi_0
xz=5.2.5=h642e427_1
yaml=0.2.5=h3422bc3_2
zeromq=4.3.4=hbdafb3b_1
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=hee7b306_1013
